requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.8
statsmodels>=0.14
PyYAML>=6.0
torch>=2.1; platform_system!="Darwin" or platform_machine!="arm64"
# Pour Apple Silicon, installez PyTorch selon les instructions officielles si besoin.
